% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ltrccf.R
\name{predict.ltrccf}
\alias{predict.ltrccf}
\title{Compute a Survival Curve from a LTRCCF model}
\usage{
\method{predict}{ltrccf}(
  object,
  newdata = NULL,
  newdata.id,
  OOB = FALSE,
  time.eval,
  time.tau = NULL
)
}
\arguments{
\item{object}{an object as returned by \code{\link{ltrccf}}.}

\item{newdata}{an optional data frame containing the test data
(with the names of the variables the same as those in \code{data} from \code{object}).}

\item{newdata.id}{optional variable name of subject identifiers for \code{newdata}.
If this is present, it will be search for in the \code{newdata} data frame.
Each group of rows in \code{newdata} with the same subject \code{id} represents
the covariate path through time of a single subject, and the result will
contain one curve per subject. If it is not specified, then predictions are
returned for each row of \code{newdata}.}

\item{OOB}{a logical specifying whether out-of-bag predictions are desired


(only if \code{newdata = NULL}).}

\item{time.eval}{a vector of time points, at which the estimated survival probabilities
will be computed.}

\item{time.tau}{an optional vector, with the \emph{i}-th entry giving the upper time limit for the
computed survival probabilities for the \emph{i}-th data of interest (i.e., only compute
survival probabilies at \code{time.eval[time.eval <= time.tau[i]]} for the \emph{i}-th
data of interest). If \code{OOB = TRUE}, the length of \code{time.tau} is equal to the length of
\code{data} used to train the \code{object};
If \code{OOB = FALSE}, the length of \code{time.tau} is equal to the length
of \code{newdata}, or equal to the length of \code{data} if \code{newdata} is not given.
The default \code{NULL} is simply to set all entries of \code{time.tau} equal to the maximum
value of \code{time.eval}, therefore all estimated survival are computed at the
same \code{time.eval}.}
}
\value{
A list containing:
   \item{survival.obj}{an object of class \code{\link[survival]{Surv}}.}
   \item{survival.probs}{the estimated survival probabilities for each data of interest.
   It is a list if the length of the estimated values differs from one to another;
   otherwise, it is a matrix with the number of columns equal to the number of the data
   of interest, number of rows equal to the number of the time points at which the estimated
   survival probabilities are computed.}
   \item{survival.tau}{the input value \code{time.tau}.}
   \item{survival.times}{the input value \code{time.eval}. }
}
\description{
Constructs a monotone nonincreasing estimated survival curve from a LTRCCF model
for any given (left-truncated) right-censored survival data with time-varying covariates.
It can also compute survival function estimates for left-truncated right-censored data
with time-invariant covariates.
}
\examples{
#### Example with data pbcsample
Formula <- Surv(Start, Stop, Event) ~ age + alk.phos + ast + chol + edema
## Fit an LTRC conditional inference forest on time-varying data
LTRCCFobj <- ltrccf(formula = Formula, data = pbcsample, id = ID,
                    mtry = 3, ntree = 50L)


## Construct an estimated survival estimate for the second subject
tpnt <- seq(0, max(pbcsample$Stop), length.out = 500)
newData <- pbcsample[pbcsample$ID == 2,]
Pred <- predict(object = LTRCCFobj, newdata = newData, newdata.id = ID,
                time.eval = tpnt)
## Since time.tau = NULL, Pred$survival.probs is in the matrix format, with dimensions:
dim(Pred$survival.probs) # length(time.eval) x nrow(newdata)
## Plot the estimated survival curve
plot(Pred$survival.times, Pred$survival.probs, type = "l", col = "red",
     xlab = "Time", ylab = "Survival probabilities")


## When time.tau is specified and some entries are different from the others
Pred2 = predict(object = LTRCCFobj, newdata = pbcsample, newdata.id = ID,
                time.eval = tpnt, time.tau = seq(100, 400,
                                                 length.out = length(unique(pbcsample$ID))))
## Then Pred2$survival.probs is a list:
class(Pred2$survival.probs)
## Plot the estimated survival curve for the subject with id = 18
plot(Pred2$survival.times[Pred2$survival.times <= Pred2$survival.tau[18]],
     Pred2$survival.probs[[18]], type = "l", col = "red",
     xlab = "Time", ylab = "Survival probabilities")


## Out-of-bag prediction
PredOOB <- predict.ltrccf(object = LTRCCFobj, OOB = TRUE, time.eval = tpnt)
}
\seealso{
\code{\link{sbrier_ltrc}} for evaluation of model fit
}
