% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ltrccf.R
\name{print.ltrccf}
\alias{print.ltrccf}
\title{Print Summary Output of a ltrccf object}
\usage{
\method{print}{ltrccf}(x)
}
\arguments{
\item{x}{an object of class \code{(ltrccf, grow)}.}
}
\description{
Print summary output after a LTRCCF model is built.
This is the default print method for objects in the class of \code{\link{ltrcrsf}}.
}
\examples{
Formula = Surv(Start, Stop, Event) ~ age + alk.phos + ast + chol + edema
# Built a LTRCCF forest on the time-varying data by specifying id, with mtry specified:
LTRCCFobj = ltrccf(formula = Formula, data = pbcsample, id = ID, mtry = 3, ntree = 50L)
print(LTRCRSFobj)
}
\seealso{
\code{\link{ltrccf}}
#' @export
}
