% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ltrcrsf.R
\name{print.ltrcrsf}
\alias{print.ltrcrsf}
\title{Print Summary Output of a ltrcrsf object}
\usage{
\method{print}{ltrcrsf}(x)
}
\arguments{
\item{x}{an object of class \code{(ltrcrsf, grow)}.}
}
\description{
Print summary output after a LTRCRSF model is built.
This is the default print method for objects in the class of \code{\link{ltrcrsf}}.
}
\examples{
#' Formula = Surv(Start, Stop, Event) ~ age + alk.phos + ast + chol + edema
# Built a LTRCRSF forest (based on bootstrapping subjects without replacement)
# on the time-varying data by specifying id, with mtry specified:
LTRCRSFobj = ltrcrsf(formula = Formula, data = pbcsample, id = ID, mtry = 3, ntree = 50L)
print(LTRCRSFobj)
}
\seealso{
\code{\link{ltrcrsf}}
}
