% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.ltrccf.R
\name{tune.ltrccf}
\alias{tune.ltrccf}
\title{Tune \code{mtry} to the optimal value with respect to out-of-bag error for a LTRCCF model}
\usage{
tune.ltrccf(
  formula,
  data,
  id,
  mtryStart = NULL,
  stepFactor = 2,
  ntreeTry = 100L,
  bootstrap = c("by.sub", "by.root", "none", "by.user"),
  samp = NULL,
  perturb = list(replace = FALSE, fraction = 0.632),
  na.action = na.pass,
  time.eval = NULL,
  time.tau = NULL,
  trace = TRUE,
  plot = FALSE,
  doBest = FALSE,
  applyfun = NULL,
  cores = NULL,
  control = partykit::ctree_control(teststat = "quad", testtype = "Univ", mincriterion
    = 0, saveinfo = FALSE, minsplit = max(ceiling(sqrt(nrow(data))), 20), minbucket =
    max(ceiling(sqrt(nrow(data))), 7), minprob = 0.01)
)
}
\arguments{
\item{formula}{a formula object, with the response being a \code{\link[survival]{Surv}}
object, with form


\code{Surv(tleft, tright, event)}.}

\item{data}{a data frame containing \code{n} rows of
left-truncated right-censored observations.}

\item{id}{variable name of subject identifiers. If this is present, it will be
searched for in the \code{data} data frame. Each group of rows in \code{data}
with the same subject \code{id} represents the covariate path through time of
a single subject. If not specified, the algorithm then assumes \code{data}
contains left-truncated and right-censored survival data with time-invariant
covariates.}

\item{mtryStart}{starting value of \code{mtry}; default is \code{sqrt(nvar)}.}

\item{stepFactor}{at each iteration, \code{mtry} is inflated (or deflated)
by this value. The default value is \code{2}.}

\item{ntreeTry}{number of trees used at the tuning step.}

\item{bootstrap}{bootstrap protocol.
(1) If \code{id} is present,
the choices are: \code{"by.sub"} (by default) which bootstraps subjects,
\code{"by.root"} which bootstraps pseudo-subjects.
Both can be with or without replacement (by default sampling is without
replacement; see the option \code{perturb} below);
(2) If \code{id} is not specified, it bootstraps the \code{data} by
sampling with or without replacement.
Regardless of the presence of \code{id}, if \code{"none"} is chosen, the
\code{data} is not bootstrapped at all. If \code{"by.user"} is choosen,
the bootstrap specified by \code{samp} is used.}

\item{samp}{Bootstrap specification when \code{bootstype = "by.user"}.
Array of dim \code{n x ntree} specifying how many times each record appears
inbag in the bootstrap for each tree.}

\item{perturb}{a list with arguments \code{replace} and \code{fraction} determining which
type of resampling, with \code{replace = TRUE} referring to the \emph{n}-out-of-\emph{n}
bootstrap and \code{replace = FALSE} referring to sample splitting. \code{fraction} is
the proportion of observations to draw without replacement.}

\item{na.action}{a function which indicates what should happen when the data contain
missing values.}

\item{time.eval}{a vector of time points, at which the estimated survival probabilities
are evaluated.}

\item{time.tau}{an optional vector, with the \emph{i}-th entry giving the upper time limit for the
computed survival probabilities for the \emph{i}-th data (i.e., only compute
survival probabilies at \code{time.eval[time.eval <= time.tau[i]]} for the \emph{i}-th
data of interest).}

\item{trace}{whether to print the progress of the search. \code{trace = TRUE} is set by default.}

\item{plot}{whether to plot the out-of-bag error as a function of \code{mtry}.}

\item{doBest}{whether to run a \code{\link{ltrccf}} object using the optimal \code{mtry} found.}

\item{applyfun}{an optional \code{lapply}-style function with arguments
\code{function(X, FUN, ...)}.
It is used for computing the variable selection criterion. The default is to use the
basic \code{lapply} function unless the \code{cores} argument is specified (see below).
See \code{\link[partykit]{ctree_control}}.}

\item{cores}{numeric. If set to an integer the \code{applyfun} is set to
\code{\link[parallel]{mclapply}} with the desired number of cores.
See \code{\link[partykit]{ctree_control}}.}

\item{control}{a list with control parameters, see \code{\link[partykit]{cforest}}.
The default values correspond to those of the default values used by \code{\link{ltrccf}}.}
}
\value{
If \code{doBest = FALSE} (default), this returns the optimal mtry value of those searched.

If \code{doBest = TRUE}, this returns the \code{\link{ltrccf}} object produced with the optimal \code{mtry}.
}
\description{
Starting with the default value of \code{mtry}, search for the optimal value
(with respect to Out-of-Bag error estimate) of \code{mtry} for \code{\link{ltrccf}}.
}
\examples{
### Example with data pbcsample
Formula = Surv(Start, Stop, Event) ~ age + alk.phos + ast + chol + edema
## mtry tuned by the OOB procedure with stepFactor 3, number of trees built 50.
mtryT = tune.ltrccf(formula = Formula, data = pbcsample, id = ID, stepFactor = 3,
                    ntreeTry = 50L, plot = TRUE)


}
\seealso{
\code{\link{sbrier_ltrc}} for evaluation of model fit when searching
for the optimal value of \code{mtry}.
}
\keyword{brier}
\keyword{errors,}
\keyword{mtry,}
\keyword{out-of-bag}
\keyword{score}
