% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ltrcrsf.R
\name{predict.ltrcrsf}
\alias{predict.ltrcrsf}
\title{Compute a Survival Curve from a LTRCRSF model}
\usage{
\method{predict}{ltrcrsf}(
  object,
  newdata = NULL,
  newdata.id,
  OOB = FALSE,
  time.eval,
  time.tau = NULL
)
}
\arguments{
\item{object}{an object as returned by \code{\link{ltrcrsf}}.}

\item{newdata}{an optional data frame containing the test data
(with the names of the variables the same as those in \code{traindata}).}

\item{newdata.id}{optional variable name of subject identifiers for \code{newdata}.
If this is present, it will be search for in the \code{newdata} data frame.
Each group of rows in \code{newdata} with the same subject \code{id} represents
the covariate path through time of a single subject, and the result will
contain one curve per subject. If it is not specified, then predictions are
returned for each row of \code{newdata}.}

\item{OOB}{a logical specifying whether out-of-bag predictions are desired


(only if \code{newdata = NULL}).}

\item{time.eval}{a vector of time points, at which the estimated survival probabilities
are computed.}

\item{time.tau}{an optional vector, with the \emph{i}-th entry giving the upper time limit for the
computed survival probabilities for the \emph{i}-th data of interest (i.e., only compute
survival probabilies at \code{time.eval[time.eval<=time.tau[i]]} for the \emph{i}-th
data of interest). If \code{OOB = TRUE}, the length of \code{time.tau} is equal to the length of
\code{data} used to train the \code{object};
If \code{OOB = FALSE}, the length of \code{time.tau} is equal to the length
of \code{newdata}, or equal to the length of \code{data} if \code{newdata} is not given.
The default \code{NULL} is simply to set all entries of \code{time.tau} equal to the maximum
value of \code{time.eval}, therefore all estimated survival are computed at the
same \code{time.eval}.}
}
\value{
A list containing:
   \item{survival.obj}{an object of class \code{\link[survival]{Surv}}.}
   \item{survival.probs}{the estimated survival probabilities for each data of interest.
   It is a list if the length of the estimated values differs from one to another;
   otherwise, it is a matrix with the number of columns equal to the number of the data
   of interest, number of rows equal to the number of the time points at which the estimated
   survival probabilities are computed.}
   \item{survival.tau}{the input value \code{time.tau}.}
   \item{survival.times}{the input value \code{time.eval}.}
}
\description{
Constructs a monotone nonincreasing estimated survival curve from a LTRCRSF model
for any given right-censored survival data with time-varying covariates. It can
also conputes survival function estimates for left-truncated right-censored data
with time-invariant covariate.
}
\seealso{
\code{\link{sbrier_ltrc}} for evaluation of model fit
}
